scripts/requirements.txt
python-jobspy==1.1.82
supabase>=2.3.0
